requirements.txt
load-dotenv==0.1.0
requests==2.31.0
requests-toolbelt==1.0.0
